MazeGame - move LimitOfViewProvider and MazeVisibilityUpdater to Painters.Map namespace
